added controller and move logic* from routes
