+++ express/app.js
@@ -14,18 +14,20 @@ nunjucks.configure('views', {
 app.set('view engine', 'nunjucks');
 // app.set('views', 'views');
 
-const adminData = require('./routes/admin');
+const adminRoutes = require('./routes/admin');
 const shopRoutes = require('./routes/shop');
 
 app.use(bodyParser.urlencoded({ extended: false }));
 app.use(express.static(path.join(__dirname, 'public')));
 
-app.use('/admin', adminData.routes);
+app.use('/admin', adminRoutes);
 app.use(shopRoutes);
 
 
 app.use((req, res, next) => {
-  res.status(404).sendFile(path.join(__dirname, 'views', '404.html'));
+  res.render('404', {
+    pageTitle: 'Page Not Found'
+  });
 });
 
 app.listen(3000, () => {
