+++ express/routes/admin.js
@@ -1,24 +1,11 @@
 const express = require('express');
-const path = require('path');
-const rootDir = require('../utils/path');
 const router = express.Router();
-
-const products = [];
-
-router.get('/add-product', (req, res, next) => {
-    res.render('add-product', {
-        pageTitle: 'Add Product',
-    });
-  });
-
-router.post('/product', (req, res, next) => {
-  products.push({ title: req.body.title });
-  
-  res.redirect(301, '/');
-});
+const productController = require('../controllers/productController');
 
 
-module.exports = {
-  routes: router,
-  products: products
-};
+router.get('/add-product', productController.getAddProduct);
+
+router.post('/product', productController.postAddProduct);
+
+
+module.exports = router;
